Move transformers to trans.py
