+++ tests/test_format.py
@@ -80,7 +80,9 @@ SOURCES = [
     "src/black/lines.py",
     "src/black/mode.py",
     "src/black/nodes.py",
+    "src/black/rusty.py",
     "src/black/strings.py",
+    "src/black/trans.py",
     "src/blackd/__init__.py",
     "src/blib2to3/pygram.py",
     "src/blib2to3/pytree.py",
